misc: Adjust some config types
Change some config types from 32 bit to 8 or 16 bit types.
Also remove some now unused ConfArray variant types.

+++ software/src/modules/em_sdcard/em_sdcard.cpp
@@ -27,9 +27,9 @@
 void EMSDcard::pre_setup()
 {
     state = Config::Object({
-        {"sd_status",       Config::Uint32(0)},
-        {"lfs_status",      Config::Uint32(0)},
-        {"card_type",       Config::Uint32(0)},
+        {"sd_status",       Config::Uint16(0)},
+        {"lfs_status",      Config::Uint16(0)},
+        {"card_type",       Config::Uint16(0)},
         {"sector_count",    Config::Uint32(0)},
         {"sector_size",     Config::Uint16(0)},
         {"manufacturer_id", Config::Uint8(0)},
