+++ software/src/modules/ws/web_sockets.cpp
@@ -568,7 +568,7 @@ void WebSockets::pre_setup() {
         {"keep_alive_peers", Config::Array({}, new Config{Config::Str("", 0, INET6_ADDRSTRLEN)}, MAX_WEB_SOCKET_CLIENTS, MAX_WEB_SOCKET_CLIENTS, Config::type_id<Config::ConfString>())},
         {"worker_active", Config::Uint8(WEBSOCKET_WORKER_DONE)},
         {"last_worker_run", Config::Uint32(0)},
-        {"queue_len", Config::Uint32(0)}
+        {"queue_len", Config::Uint16(0)}
     });
 
     Config *state_keep_alive_fds = static_cast<Config *>(state.get("keep_alive_fds"));
